examples/helloworld/greeter_client_python/requirements.txt
grpcio==1.32.0
protobuf==3.13.0
